camelot-prototype/requirements.txt
camelot-py[cv]
pandas
openpyxl
tabulate
